most renames, small fixes, introduction to views/services file management
