+++ ngApp/src/app/app.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
-import { AuthService } from './auth.service';
+import { AuthService } from './services/authorization/auth.service';
+
 
 @Component({
   selector: 'app-root',
