+++ ngApp/src/app/app.module.ts
@@ -5,14 +5,14 @@ import { HttpClientModule, HTTP_INTERCEPTORS } from "@angular/common/http";
 import { AppRoutingModule } from './app-routing.module';
 
 import { AppComponent } from './app.component';
-import { RegisterComponent } from './register/register.component';
-import { LoginComponent } from './login/login.component';
-import { EventsComponent } from './events/events.component';
-import { SpecialEventsComponent } from './special-events/special-events.component';
-import { AuthService } from './auth.service';
-import { EventService } from './event.service';
-import { AuthGuard } from './auth.guard';
-import { TokenInterceptorService} from './token-interceptor.service';
+import { RegisterComponent } from './views/register/register.component';
+import { LoginComponent } from './views/login/login.component';
+import { EventsComponent } from './views/events/events.component';
+import { SpecialEventsComponent } from './views/special-events/special-events.component';
+import { AuthService } from './services/authorization/auth.service';
+import { EventService } from './services/event/event.service';
+import { AuthGuard } from './services/authorization/auth.guard';
+import { TokenInterceptorService} from './services/token-interceptor/token-interceptor.service';
 
 @NgModule({
   declarations: [
