+++ ngApp/src/app/services/token-interceptor/token-interceptor.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, Injector } from '@angular/core';
 import { HttpInterceptor } from "@angular/common/http";
-import { AuthService } from './auth.service';
+import { AuthService } from './../authorization/auth.service';
 
 @Injectable({
   providedIn: 'root'
