+++ ngApp/src/app/views/events/events.component.html
