+++ ngApp/src/app/views/events/events.component.spec.ts
