+++ ngApp/src/app/views/events/events.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { EventService } from '../event.service';
+import { EventService } from './../../services/event/event.service';
 
 @Component({
   selector: 'app-events',
