+++ ngApp/src/app/views/login/login.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { AuthService } from '../auth.service';
+import { AuthService } from './../../services/authorization/auth.service';
 
 @Component({
   selector: 'app-login',
