+++ ngApp/src/app/views/register/register.component.scss
