+++ ngApp/src/app/views/special-events/special-events.component.html
